Reorganize values utils
