+++ pkg/values/diff_test.go
@@ -1,4 +1,4 @@
-package chart
+package values
 
 import (
 	"testing"
